Bug Fixing #5 & Uploading add_progressController
Here I have fixed some bugs in patients.php to so that data properly sent to the backend. I have also put the links and change the names of the input fields which are needed to create progress for a patient. Adding New Progress for a patient is successfully completed.